REFACTOR: cambiado un tipo de dato de id de int a long

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/JobPackageService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface JobPackageService {
 
-    JobPackage create(long postId, String title, String description, String price, int rateType);
+    JobPackage create(long postId, String title, String description, String price, long rateType);
 
     JobPackage findById(long id);
